+++ pkg/cloudprovider/providers/gce/BUILD
@@ -13,6 +13,23 @@ go_library(
     srcs = [
         "doc.go",
         "gce.go",
+        "gce_backendservice.go",
+        "gce_cert.go",
+        "gce_clusters.go",
+        "gce_disks.go",
+        "gce_firewall.go",
+        "gce_forwardingrule.go",
+        "gce_healthchecks.go",
+        "gce_instancegroup.go",
+        "gce_instances.go",
+        "gce_loadbalancer.go",
+        "gce_op.go",
+        "gce_routes.go",
+        "gce_staticip.go",
+        "gce_targetproxy.go",
+        "gce_urlmap.go",
+        "gce_util.go",
+        "gce_zones.go",
         "token_source.go",
     ],
     tags = ["automanaged"],
